Don't trim domain of of SPF with `mfrom` scope
Mail::SPF check are done using the full host (i.e. `foo.example.com` in
`Foo <bar@foo.example.com>`) as identity for the `mfrom` scope:
https://metacpan.org/dist/Mail-SPF/source/lib/Mail/SPF/Request.pm#L265-267

Currently, this plugin tells Mail::DMARC::PurePerl that the identity used for the check for the `mfrom` scope is what SpamAssassin gives when it's asked `EnvelopeFrom:domain`.

However, `:domain` gives the trimmed version of the host (i.e.
`example.com` in `Foo <bar@foo.example.com>`).

This results in alignement failing in case both the envelope sender and
the from sender have a subdomain in their host (e.g. `bar@foo.example.com` ;
it would work if it was `bar@example.com`).

This commit fixes this issue.